requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic
slowapi
email-validator
pandas
openpyxl
redis
rq
python-dotenv
